This fixes a memory leak in slpd.
Note from patch submitter: Pete
There has been no growth since starting slpd.  The fix appears to have corrected the leak.The pointer would need to be added to pNormalisedSrvtype would need to be
either freed before leaving the Register call (if not stored indexed) or cleaned up by the Deregistration.  The patch adds it to the structure to allow pNormalisedSrvtype to be freed in a consistent manner (whether indexed or
non-indexed storage database storage is done) by the Deregistration code.


git-svn-id: https://openslp.svn.sourceforge.net/svnroot/openslp/trunk@1649 3ca0f695-f50d-0410-9461-b46777378323